chore Model 분리
