+++ Projects/Features/MainFeature/Sources/ViewModels/BroadcastCollectionViewModel.swift
@@ -5,31 +5,6 @@ import BaseFeatureInterface
 import BroadcastDomainInterface
 import LiveStationDomainInterface
 
-public struct Channel: Hashable {
-    let id: String
-    let name: String
-    var thumbnailImageURLString: String
-    var thumbnailImage: UIImage?
-    let owner: String
-    let description: String
-
-    public init(
-        id: String,
-        title: String,
-        thumbnailImageURLString: String = "",
-        thumbnailImage: UIImage? = nil,
-        owner: String = "",
-        description: String = ""
-    ) {
-        self.id = id
-        self.name = title
-        self.thumbnailImageURLString = thumbnailImageURLString
-        self.thumbnailImage = thumbnailImage
-        self.owner = owner
-        self.description = description
-    }
-}
-
 public class BroadcastCollectionViewModel: ViewModel {
     public struct Input {
         let fetch: PassthroughSubject<Void, Never> = .init()
